[YOURLSClient] Cog model restructure
Restructure cog `YOURLSClient` according to SFUAnime/Ren#608. No
functional changes. Constants are moved to `constants.py`.
