+++ cogs/yourlsClient/constants.py
@@ -1,0 +1,8 @@
+KEY_API = "api"
+KEY_SIGNATURE = "signature"
+
+BASE_GUILD = {
+    KEY_API: None,
+    KEY_SIGNATURE: None,
+}
+DEFAULT_ERROR = "Something went wrong, please try again later, or check your console for details."
